refactor(tests): rename glb render helper to renderGlbToPng
