+++ tests/features/models-from-footprint-strings/footprint-string-models.test.tsx
@@ -1,6 +1,6 @@
 import { Circuit } from "tscircuit"
 import { test, expect } from "bun:test"
-import { renderGlbToPng } from "../../helpers"
+import { renderGlbToPng } from "../../renderGlbToPng"
 import { convertCircuitJsonToGltf } from "../../../lib"
 
 test("models-from-footprint-strings", async () => {
