+++ tests/features/translucent-glb-models/translucent-vs-opaque-glb-models.test.tsx
@@ -1,6 +1,6 @@
 import { Circuit } from "tscircuit"
 import { test, expect } from "bun:test"
-import { renderGlbToPng } from "../../helpers"
+import { renderGlbToPng } from "../../renderGlbToPng"
 import { convertCircuitJsonToGltf } from "../../../lib"
 
 test("translucent-vs-opaque-glb-models", async () => {
