+++ tests/renderGlbToPng.ts
