+++ tests/repro/repro12-testpoint.test.tsx
@@ -1,5 +1,5 @@
 import { expect, test } from "bun:test"
-import { renderGlbToPng } from "../helpers"
+import { renderGlbToPng } from "../renderGlbToPng"
 import { Circuit } from "tscircuit"
 import { convertCircuitJsonTo3D, convertCircuitJsonToGltf } from "../../lib"
