+++ tests/repro/repro9-stm32f100c8t6b.test.tsx
@@ -1,6 +1,6 @@
 import { Circuit } from "tscircuit"
 import { test, expect } from "bun:test"
-import { renderGlbToPng } from "../helpers"
+import { renderGlbToPng } from "../renderGlbToPng"
 import { convertCircuitJsonToGltf } from "../../lib"
 import type { ChipProps } from "@tscircuit/props"
