+++ tests/snapshot/board-outline-direction.test.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "bun:test"
-import { renderGlbToPng } from "../helpers"
+import { renderGlbToPng } from "../renderGlbToPng"
 import { convertCircuitJsonToGltf } from "../../lib/index"
 import type { CircuitJson } from "circuit-json"
